requirements.txt
playwright==1.41.0
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.0.0
numpy>=1.24.0

